new feature

+++ src/main/java/com/ndroid/shopping/shopping_api/dto/UserModelDto.java
@@ -2,16 +2,11 @@ package com.ndroid.shopping.shopping_api.dto;
 
 import java.time.LocalDateTime;
 
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 import lombok.Data;
 
 @Data
 public class UserModelDto {
-    private Integer id;
+    private Long id;
     private String username;
     private String email;
     private String phone;
